Extract findVisibleFunctions() et. al. from functionResolution.cpp
