+++ compiler/include/astutil.h
@@ -183,7 +183,4 @@ bool isTypeExpr(Expr* expr);
 Symbol* getSvecSymbol(CallExpr* call);
 void collectUsedFnSymbols(BaseAST* ast, std::set<FnSymbol*>& fnSymbols);
 
-// move to resolve when scope resolution is put in resolution directory
-BlockStmt* getVisibilityBlock(Expr* expr);
-
 #endif
