+++ compiler/resolution/wrappers.cpp
@@ -50,6 +50,7 @@
 #include "stmt.h"
 #include "stringutil.h"
 #include "symbol.h"
+#include "visibleFunctions.h"
 
 //########################################################################
 //# Static Function Forward Declarations
